requirements.txt
scapy
mysql-connector-python
cryptography
scikit-learn
numpy
